--- frontend/tailwind.config.js
@@ -1,17 +1,0 @@
-// tailwind.config.js
-const {nextui} = require("@nextui-org/theme");
-
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    './src/**/*.{js,jsx,ts,tsx}', 
-    './public/index.html',
-    "./node_modules/@nextui-org/theme/dist/components/navbar.js"
-  ],
-  theme: {
-    extend: {},
-  },
-  darkMode: "class",
-  plugins: [nextui()],
-};
-
