+++ my-react-router-app/app/routes/cs_dev/components/ProjectDetails.tsx
@@ -3,7 +3,7 @@
  * actual project.
 */
 
-import { allSkills } from "../../../../../frontend/src/data/skillsProjects";
+import { allSkills } from "../data/skillsProjects";
 
 interface ProjectDetailsProps {
     name: string;
