+++ my-react-router-app/app/routes/cs_dev/components/SocialsDisplay.tsx
@@ -1,4 +1,4 @@
-import { socialLinks, SocialLink } from "../../../../../frontend/src/data/socialLinks";
+import { socialLinks, SocialLink } from "../data/socialLinks";
 
 export default function DevSocials() {
   return (
