+++ my-react-router-app/app/routes/cs_dev/images/index.html
